requirements.txt
feedparser
aiohttp
requests
python-dotenv
